+++ core/cu29/src/lib.rs
@@ -5,6 +5,7 @@ pub mod copperlist;
 pub mod curuntime;
 pub mod cutask;
 pub mod monitoring;
+pub mod simulation;
 
 pub use config::read_configuration;
 pub use cu29_clock as clock;
